+++ proxy/test/echo_proxy_test.go
@@ -4,12 +4,14 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/suite"
 	"github.com/temporalio/s2s-proxy/config"
 	"github.com/temporalio/s2s-proxy/encryption"
 	"go.temporal.io/api/workflowservice/v1"
 	"go.temporal.io/server/api/adminservice/v1"
+	replicationpb "go.temporal.io/server/api/replication/v1"
 	"go.temporal.io/server/client/history"
 	"go.temporal.io/server/common/log"
 )
@@ -33,6 +35,13 @@ var (
 		ClusterID: 2,
 		ShardID:   4,
 	}
+
+	emptyReq = &adminservice.StreamWorkflowReplicationMessagesRequest{
+		Attributes: &adminservice.StreamWorkflowReplicationMessagesRequest_SyncReplicationState{
+			SyncReplicationState: &replicationpb.SyncReplicationState{
+				HighPriorityState: &replicationpb.ReplicationState{},
+			},
+		}}
 )
 
 type (
@@ -592,4 +601,46 @@ func (s *proxyTestSuite) Test_Echo_WithMuxTransport() {
 
 	s.NoError(err)
 	s.Equal("EchoServer", r.ClusterName)
+}
+
+func (s *proxyTestSuite) Test_ForceStopSourceServer() {
+	logger := log.NewTestLogger()
+
+	echoServerInfo := clusterInfo{
+		serverAddress:  echoServerAddress,
+		clusterShardID: serverClusterShard,
+	}
+
+	echoClientInfo := clusterInfo{
+		serverAddress:  echoClientAddress,
+		clusterShardID: clientClusterShard,
+		s2sProxyConfig: createEchoClientConfig(),
+	}
+
+	echoServer := newEchoServer(echoServerInfo, echoClientInfo, "EchoServer", logger, nil)
+	echoClient := newEchoServer(echoClientInfo, echoServerInfo, "EchoClient", logger, nil)
+
+	echoServer.start()
+	echoClient.start()
+
+	stream, err := echoClient.CreateStreamClient()
+	s.NoError(err)
+	_, err = sendRecv(stream, []int64{1})
+	s.NoError(err)
+
+	echoServer.server.ForceStop()
+
+	// ForceStop cause sourceStreamClient.Recv in Upstream loop within
+	// StreamWorkflowReplicationMessages handler to fail. Wait for
+	// StreamWorkflowReplicationMessages handler returns, which stop
+	// Downstream loop.
+	time.Sleep(time.Second)
+
+	err = stream.Send(emptyReq)
+
+	// This should fail because StreamWorkflowReplicationMessages handler stopped.
+	s.ErrorContains(err, "EOF")
+
+	stream.CloseSend()
+	echoClient.stop()
 }
